+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/localstorage/LocalFileLoader.java
@@ -22,31 +22,20 @@
 package com.microsoft.applicationinsights.agent.internal.localstorage;
 
 import com.microsoft.applicationinsights.agent.internal.common.OperationLogger;
-import com.microsoft.applicationinsights.agent.internal.common.ThreadPoolUtils;
 import java.io.File;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.file.Files;
-import java.util.Collection;
-import java.util.Date;
 import java.util.Queue;
 import java.util.concurrent.ConcurrentLinkedDeque;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 import javax.annotation.Nullable;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
-import org.checkerframework.checker.nullness.qual.Nullable;
 
 /** This class manages loading a list of {@link ByteBuffer} from the disk. */
 public class LocalFileLoader {
 
   private static final String TEMPORARY_FILE_EXTENSION = ".tmp";
-
-  private static final ScheduledExecutorService scheduledExecutor =
-      Executors.newSingleThreadScheduledExecutor(
-          ThreadPoolUtils.createDaemonThreadFactory(LocalFileLoader.class));
 
   private final LocalFileCache localFileCache;
   private final File telemetryFolder;
@@ -56,26 +45,11 @@ public class LocalFileLoader {
 
   private final Queue<File> toBeDeletedFileQueue = new ConcurrentLinkedDeque<>();
 
-  public LocalFileLoader(LocalFileCache localFileCache, File telemetryFolder) {
-    this(localFileCache, telemetryFolder, null, null);
-  }
-
-  // purgeIntervalSeconds and expiredIntervalSeconds are used by tests only
   public LocalFileLoader(
       LocalFileCache localFileCache,
-      File telemetryFolder,
-      @Nullable Long purgeIntervalSeconds,
-      @Nullable Long expiredIntervalSeconds) {
+      File telemetryFolder) {
     this.localFileCache = localFileCache;
     this.telemetryFolder = telemetryFolder;
-
-    // run purge task daily to clean up expired files that are older than 48 hours.
-    long interval =
-        purgeIntervalSeconds == null ? TimeUnit.DAYS.toSeconds(1) : purgeIntervalSeconds;
-    long expiredInterval =
-        expiredIntervalSeconds == null ? TimeUnit.DAYS.toSeconds(2) : expiredIntervalSeconds;
-    scheduledExecutor.scheduleWithFixedDelay(
-        new PurgeExpiredFilesTask(expiredInterval), interval, interval, TimeUnit.SECONDS);
   }
 
   // Load ByteBuffer from persisted files on disk in FIFO order.
@@ -156,65 +130,10 @@ public class LocalFileLoader {
     } catch (IOException ex) {
       // TODO (heya) track file deletion failure via Statsbeat
       operationLogger.recordFailure("Fail to delete " + file.getName(), ex);
-      retryDelete(file);
+      LocalStorageUtils.retryDelete(file);
     } catch (SecurityException ex) {
       operationLogger.recordFailure(
           "Unable to delete " + file.getName() + ". Access is denied.", ex);
-    }
-  }
-
-  // retry delete 3 times when it fails.
-  private static void retryDelete(File file) {
-    if (file.exists()) {
-      for (int i = 0; i < 3; i++) {
-        try {
-          Thread.sleep(500);
-          System.gc();
-          if (file.delete()) {
-            break;
-          }
-        } catch (InterruptedException ex) {
-          operationLogger.recordFailure("Fail to perform Thread.sleep.", ex);
-        }
-      }
-    }
-  }
-
-  private class PurgeExpiredFilesTask implements Runnable {
-
-    private final long expiredIntervalSeconds;
-
-    private PurgeExpiredFilesTask(long expiredIntervalSeconds) {
-      this.expiredIntervalSeconds = expiredIntervalSeconds;
-    }
-
-    @Override
-    public void run() {
-      purgedExpiredFiles();
-    }
-
-    private void purgedExpiredFiles() {
-      Collection<File> files = FileUtils.listFiles(telemetryFolder, new String[] {"trn"}, false);
-      for (File file : files) {
-        if (expired(file.getName())) {
-          try {
-            Files.delete(file.toPath());
-          } catch (IOException ex) {
-            operationLogger.recordFailure("Fail to delete the expired " + file.getName(), ex);
-            retryDelete(file);
-          }
-        }
-      }
-    }
-
-    // files that are older than expiredIntervalSeconds (default 48 hours) are expired and need to
-    // be deleted permanently.
-    private boolean expired(String fileName) {
-      String time = fileName.substring(0, fileName.lastIndexOf('-'));
-      long milliseconds = Long.parseLong(time);
-      Date twoDaysAgo = new Date(System.currentTimeMillis() - 1000 * expiredIntervalSeconds);
-      Date fileDate = new Date(milliseconds);
-      return fileDate.before(twoDaysAgo);
     }
   }
 }
