+++ agent/agent-tooling/src/main/java/com/microsoft/applicationinsights/agent/internal/localstorage/LocalFileSender.java
@@ -61,6 +61,7 @@ public class LocalFileSender implements Runnable {
       ByteBuffer buffer = localFileLoader.loadTelemetriesFromDisk();
       if (buffer != null) {
         CompletableResultCode resultCode = telemetryChannel.sendRawBytes(buffer);
+        resultCode.join(5, TimeUnit.SECONDS); // wait for request to be completed.
         localFileLoader.updateProcessedFileStatus(resultCode.isSuccess());
       }
     } catch (RuntimeException ex) {
